+++ src/lib/types/variable.ts
@@ -3,3 +3,10 @@ export type Variable = {
   key: string;
   value: string;
 };
+
+export type VariableMatch = {
+  name: string;
+  start: number;
+  end: number;
+  raw: string;
+};
